Update README and Course Content, Modify File Structure
Updated README file to better explain what is and what isn't included in this version of the course.

Updated course content to include materials covered up to the middle of Week 5.

Modified the file structure so each week's course structure is in a single markdown file and not in a folder.